Re-organization parsec-cli
- Move sub-command to `commands` mod.
- Create `testenv_utils` module to handle test environment setup (shared with `run_testenv` sub-command).
