+++ cli/src/commands/claim_invitation.rs
