+++ cli/src/commands/invite_device.rs
