+++ cli/src/commands/list_devices.rs
