+++ cli/src/commands/list_invitations.rs
