+++ cli/src/commands/list_users.rs
