+++ cli/src/commands/list_workspaces.rs
