+++ cli/src/commands/ls.rs
