+++ cli/src/commands/remove_device.rs
