+++ cli/src/commands/shamir_setup.rs
