+++ cli/src/commands/share_workspace.rs
