+++ cli/src/commands/stats_server.rs
