+++ cli/src/commands/status_organization.rs
